backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cachetools==6.2.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==48.0.0
dateparser==1.4.0
deprecation==2.1.0
distro==1.9.0
fastapi==0.136.1
fsspec==2026.4.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.15
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.3.0
langchain-core==1.4.0
langchain-openai==1.2.1
langchain-protocol==0.0.15
langgraph==1.2.0
langgraph-checkpoint==4.1.0
langgraph-prebuilt==1.1.0
langgraph-sdk==0.3.14
langsmith==0.8.3
markdown-it-py==4.2.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
openai==2.36.0
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
postgrest==2.30.0
propcache==0.5.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
PyYAML==6.0.3
realtime==2.30.0
regex==2026.5.9
requests==2.34.0
requests-toolbelt==1.0.0
rich==14.3.4
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
storage3==2.30.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.30.0
supabase-auth==2.30.0
supabase-functions==2.30.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.7.0
uuid_utils==0.15.0
uvicorn==0.46.0
websockets==15.0.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
